Change init.d invocation to use invoke-rc.d
This is the correct way to restart services from within a `postinst`: https://www.debian.org/doc/debian-policy/ch-opersys.html#s9.3.3.2

Invoke-rc.d takes into account the current runlevel, which is important for various reasons, one of which is creating Docker images. There is also a bug in the tomcat7 init.d script (or somewhere in this chain), which manages to fail within a Docker environment, so the building of a Docker image fails.